requirements.txt
lmdb==1.7.5
numpy==2.4.2
openai==2.18.0
opencv_python_headless==4.11.0.86
Pillow==12.1.0
pydantic==2.12.5
tqdm==4.67.1
mem0ai
faiss-cpu
